requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
idna==3.4
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
